cli/apply: Clean up Deno and Node modes
Make the code more readable by moving Deno and Node mode apply logic
into their own files.

No functional changes, just code movement.
